# internal/assets/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tumble API",
    "description": "API for the Tumble content application. The v1 API provides a RESTful, versioned interface with consistent request/response formats. Legacy endpoints are deprecated and will be removed in a future release.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for protected endpoints"
      }
    },
    "schemas": {
      "APIError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code",
                "enum": ["bad_request", "unauthorized", "forbidden", "not_found", "validation_error", "internal_error"]
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "description": "Field-level validation errors (for validation_error code)"
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"],
        "example": {
          "error": {
            "code": "not_found",
            "message": "Link 42 does not exist"
          }
        }
      },
      "APIMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "description": "Total number of items available"
          },
          "limit": {
            "type": "integer",
            "description": "Maximum items per page"
          },
          "offset": {
            "type": "integer",
            "description": "Number of items skipped"
          }
        },
        "required": ["total", "limit", "offset"]
      },
...
      "APIQuoteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique quote ID"
          },
          "quote": {
            "type": "string",
            "description": "The quote text"
          },
          "author": {
            "type": "string",
            "description": "Author of the quote"
          },
          "poster": {
            "type": "string",
            "description": "Username who submitted the quote"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the quote was submitted"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tags associated with this quote"
          }
        },
        "required": ["id", "quote", "created_at"],
        "example": {
          "id": 42,
          "quote": "The only way to do great work is to love what you do.",
          "author": "Steve Jobs",
          "poster": "alice",
          "created_at": "2026-01-15T10:30:00Z",
          "tags": ["inspiration", "tech"]
        }
      },
      "APIQuotesListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/APIQuoteResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/APIMeta"
          }
        },
        "required": ["data", "meta"]
      },
...
      "APITagResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique tag ID"
          },
          "tag": {
            "type": "string",
            "description": "Tag string (lowercased, no spaces)"
          },
          "resource_type": {
            "type": "string",
            "enum": ["link", "quote"],
            "description": "Type of resource this tag is attached to"
          },
          "resource_id": {
            "type": "integer",
            "description": "ID of the resource this tag is attached to"
          },
          "created_by": {
            "type": "string",
            "description": "Username who created this tag"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the tag was created"
          }
        },
        "required": ["id", "tag", "resource_type", "resource_id", "created_at"],
        "example": {
          "id": 1,
          "tag": "golang",
          "resource_type": "link",
          "resource_id": 42,
          "created_by": "alice",
          "created_at": "2026-02-08T14:30:00Z"
        }
      },
      "APITagsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/APITagResponse"
            }
          }
        },
        "required": ["data"]
      },
      "TagCreateRequest": {
        "type": "object",
        "properties": {
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of tag strings (lowercased, no spaces, hyphens/underscores allowed)"
          },
          "user": {
            "type": "string",
            "description": "Username creating the tags (required for authorization if not using API key)"
          }
        },
        "required": ["tags"],
        "example": {
          "tags": ["golang", "web-dev", "tutorial"],
          "user": "alice"
        }
      },
      "QuoteCreateRequest": {
        "type": "object",
        "properties": {
          "quote": {
            "type": "string",
            "description": "The quote text"
          },
          "author": {
            "type": "string",
            "description": "Author of the quote"
          },
          "poster": {
            "type": "string",
            "description": "Username of the submitter"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional tags to add to the quote (lowercased, no spaces)"
          }
        },
        "required": ["quote"],
        "example": {
          "quote": "The only way to do great work is to love what you do.",
          "author": "Steve Jobs",
          "poster": "alice",
          "tags": ["inspiration"]
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/quotes": {
      "get": {
        "summary": "List Quotes",
        "description": "Returns a paginated list of quotes.",
        "tags": ["Quotes"],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results (default: 50, max: 1000)",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip (default: 0)",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of quotes with pagination metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIQuotesListResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "description": "Plain text representation of quotes"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Quote",
        "description": "Submits a new quote. Only the quote text is required.",
        "tags": ["Quotes"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quote created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIQuoteResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "42",
                  "description": "Quote ID (plain text format)"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (malformed JSON)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (missing quote text)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotes/{id}": {
      "get": {
        "summary": "Get Quote",
        "description": "Returns a single quote by ID.",
        "tags": ["Quotes"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The quote ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quote details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIQuoteResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string",
                  "description": "Plain text representation"
                }
              }
            }
          },
          "404": {
            "description": "Quote not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Quote",
        "description": "Deletes a quote by ID. Requires API key authentication.",
        "tags": ["Quotes"],
        "security": [
          { "apiKey": [] }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The quote ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Quote deleted successfully"
          },
          "401": {
            "description": "Unauthorized (missing API key)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (invalid API key)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Quote not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/quotes/{id}/tags": {
      "get": {
        "summary": "List Tags for Quote",
        "description": "Returns all tags associated with a quote.",
        "tags": ["Tags"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The quote ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tags for the quote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APITagsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add Tags to Quote",
        "description": "Adds tags to a quote. Authorized if admin API key is present, or if within 10 minutes of quote creation and username matches the original author.",
        "tags": ["Tags"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The quote ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tags created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APITagsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Quote not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (invalid tag format)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Replace Tags on Quote",
        "description": "Replaces all tags on a quote. Same authorization rules as POST.",
        "tags": ["Tags"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The quote ID",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tags replaced successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APITagsResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "404": {
            "description": "Quote not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          }
        }
      }
    },
...
    "/quote/": {
      "get": {
        "summary": "Get a Random Quote (Deprecated)",
        "description": "DEPRECATED: Use GET /api/v1/quotes instead.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "responses": {
          "200": {
            "description": "Random Quote"
          }
        }
      },
      "post": {
        "summary": "Submit a Quote (Deprecated)",
        "description": "DEPRECATED: Use POST /api/v1/quotes instead.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "quote": { "type": "string" },
                  "author": { "type": "string" },
                  "poster": { "type": "string" }
                },
                "required": ["quote"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Quote created"
          },
          "400": {
            "description": "Missing inputs"
          }
        }
      }
    },
    "/quote/{id}": {
      "get": {
        "summary": "Get Quote by ID (Deprecated)",
        "description": "DEPRECATED: Use GET /api/v1/quotes/{id} instead.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the quote",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quote found"
          },
          "404": {
            "description": "Quote not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a quote (Deprecated)",
        "description": "DEPRECATED: Use DELETE /api/v1/quotes/{id} instead.",
        "deprecated": true,
        "tags": ["Deprecated"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the quote",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Quote deleted"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
...
}